changed folder names and cleaned up repo
