--- README.md
@@ -1,34 +1,0 @@
-WegrzynLab
-==========
-
-find5.py:
-
-	Script to find the largest 5 contigs in a FASTA file. To 	use, run the script from command prompt and put in the 	exact	filename of the FASTA file when prompted.
-
-find5After.py:
-
-	As above. This version is just to allow using the script 	on multiple sources in the same directory.
-
-completeFinder.py:
-
-	Script to find contigs labelled as complete in a FASTA 	file. Usage is the same as the previous two scripts.
-
-fastaGrabber.py
-
-	Skeleton code to grab FASTA files contigs
-
-trinityGeneFinder.py
-
-	Script to find the longest sequences for each gene in a 	FASTA file (genes determined by c#)
-
-chloroplastFinder.py
-
-	Script to find genes labelled as Chloroplasts from the 	job.py output
-
-interproAppender.py
-
-	Script to append interpro information into annotation file
-
-fastaSorter.py
-
-	Script to seperate sequences of input file into those 	above and below a threshold set by the user
