+++ enTAP/clear_dir.sh
@@ -1,0 +1,2 @@
+rm -rf blastxml* default_output* log* nohits* uninformative* contam* shell* test_* combined_annotation*
+
