--- fastaGrabber.py
@@ -1,53 +1,0 @@
-## Script Name: fastaGrabber.py
-## Purpose: Provide code skeleton for FASTA format readers
-##
-## Created by James Pickett
-## University of Connecticut
-##
-## Version: 1.0.0
-## Last Edit 7/16/2014
-## Usage: Run from terminal, filename can be included in line or a query will appear on run (Wildcard and Relative OK)
-
-import glob
-import sys
-
-cmdInput = sys.argv[len(sys.argv) - 1]
-target = glob.glob(cmdInput)[0]
-if(cmdInput == sys.argv[0]):
-	target = glob.glob(raw_input("Enter the name of the file to be parsed (Relative and wildcards OK) \n"))[0]
-
-geneSeq = open(target,'r')  #Retrieve file contents
-#storageFile = open('FiveLongestSequences.txt','w')     ##Commented out to avoid needless runs,  uncomment and change filename and print statements for full functionality
-
-pos = 0
-
-name = ''
-sequence = ''
-
-
-geneSeqString = geneSeq.read() #Store file contents as a single string
-
-
-
-while True:
-
-
-	nameStart = geneSeqString.find('>', pos) #Stores the index of the first '>' character in nameStart. '>' signifies the start of a FASTA-format sequence, 'pos' indicates the position to start searching from, and is 0 at first run
-
-	if(nameStart == -1): #Python returns a -1 if the character is not found in the string, indicating there are no more occurences and running again is unnecessary
-		break
-
-	nameEnd = geneSeqString.find('\n', nameStart) #Finds the first newline character after the index stored above, which indicates the end of the FASTA sequence description and beginning of sequence data
-
-	name = geneSeqString[nameStart:nameEnd] #Stores the entire sequence desription as name
-	sequence = geneSeqString[nameEnd + 1:geneSeqString.find('>',nameEnd) - 1] #Stores the sequence
-
-
-	if(nameStart == -1):
-		sequence = geneSeqString[geneSeqString.find('\n',pos) + 1:-1] #If there are no more sequences, the rest of the file is the current sequence
-
-	pos = nameStart + 1
-
-# for x in range(1,6): #Writes stored sequences to previously specified file       ##Comemented out for same reason as storageFile line
-# 	storageFile.write("     ")
-# 	storageFile.write("\n")
